requirements.txt
khl.py
bilibili-api-python
# pip install khl.py
# pip install playwright
# python -m playwright install